app: synchrona: misc improvements
This change is mainly about improving how 'jif_config' is constructed in
order to be passed to pyadi-dt. We had three loops in the process of
doing that while one loop is more enough. Now, everything is being done
inside 'create_hmc7044_clock_config()' which makes sense.

The rest of the changes are more about some new lines where needed and
some code reordering. Also note that commented code was removed. Let's
just add it when we are ready to do so.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>