app: synchrona: set meaningful names on hmc7044 'adi,extended-name'
This property is being set depending on the keys of the clock config
dictionaries. Before, just a number was being set that did not even
correlate to the mapping between hmc7044 and the GUI channels. With this
change "HMC7044_CLKOUTXX_CHYY" will be used and XX and YY will be so that
it corresponds to the following mapping (device - GUI):

0 - CH11;
1 - CH12;
2 - CH14;
3 - CH13;
4 - CH8;
5 - CH10;
6 - CH6;
7 - CH4;
8 - CH1;
9 - CH2;
10 - CH3;
11 - CH5;
12 - CH9;
13 - CH7.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>