installers: common: refactor systemd service
This does some refactoring of the systemd service:

* No need to 'touch' the service file;
* Explicitly remove the rpi-ad9545-hmc7044 overlay. 'dtoverlay -r'
removes the last overlay in the list of loaded overlays. While this
should not be a problem for us since rpi-ad9545-hmc7044 should the
only one in the list, better to be explict.
* Add stop steps so that if we stop the service, we unbind everything
done by starting it; Note that we do not remove the overlay otherwise
we would not be able to poweroff;
* Don't call the rebind.sh script. In fact this was problematic and
could, sometimes, lead to a race in the kernel. Note that when
remove/add overlays, this will already bind/unbind the devices. When we
were adding the overlay, the added devices start to bind and it could
happen that the script would see that the ad9545 device was already
binded while the hmc7044 was not (but was in fact in the process and did
it). Hence, we would only unbind the ad9545 driver leaving the hmc7044
in a very bad state. We don't need this script and the goal is to remove
it...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>