installers: remove rpi-ad9545-hmc7044.dtbo
Having a copy of the overlay devicetree here is just maintenance burden
as we need to make sure it stays in sync with whatever we have in the
kuiper image.

Hence, we just assume that the overlay in the image is the
default/base devicetree to use and copy it into our installation path
during installation. When uninstalling we need to restore it in the system.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>