app: synchrona: remove set_vcxo_pin.sh script
There's no point in having this script as we can directly write in the
sysfs file. By the time the app is running, the gpio was already
exported by the systemd service and the pins is properly configured
though the overlay devicetree. Hence we should only care to set the
proper value in the gpio.

As a side effect, the code also looks neater...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>