treewide: remove dependency on installation path
By default things are installed under "/var/www/html/". However, the
user is free to change it. Hence we should not rely on it and we should
not directly call things that depend on that path. Thus the
reload_dtb.sh script is now being copied into /etc/raspap/synchrona
(as the user has no way to change it) and the app is being changed
accordingly. Moreover, we also need to take this into account when
creating the systemd service and so we need to use the current
installation path and not an hardcode path.

Note that to keep things consistent, reload_dtb is now under installers/

Signed-off-by: Nuno Sá <nuno.sa@analog.com>