includes: synchrona: call reload_dtb script
We already have a script that is used in the python app to reload the
devicetree overlay. So, let's be consistent and also call it here. With
this change we also need to update the sudoers file so that we can run
the script as sudo (and hence remove the dtoverlay rules).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>