synchrona: support reporting active reference
The reference being used might be different independent of the input
priorities selected by the user. Hence, it makes sense to display the
actual reference being used in the GUI. On top of this, a LED is also
set (close to the reference box) so that if it's green the PLLs are
locked (and all should be working) and red if the PLLs are unlocked.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>