app: python: synchrona: retrieve error string
In case we cannot calculate the hmc7044 clock configuration, return an
errno string clearly stating that. So far, this was the only place where
some kind of error information was returned to the GUI. Hence, let's
fist change this behavior. A future patch will add more changes so that
we can, hopefully, catch all the possible errors and retrieve them.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>