app: python: synchrona: set vcxo gpio as 'output-low'
When the value of the vcxo gpio is set to 1, using the pull-down
property is not enough to make it 0. This work as expected because the
gpio is being exported and the value is directly written depending on
the desired vcxo.

However, in a following change, exporting the gpio will be removed since
that should not be needed (as long as all the pinctrl devicetree
properties are properly set).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>