app: python: synchrona: improve error handling/report
This change improves the error handling and report. It adds a new
simple exception class to mainly distinguish from errors that can be
recoverable (invalid set of frequencies sent by the user) from errors
that should not happen and are likely bugs.

Now, all errors/exceptions that occur in the configuration API should be
catched and a proper message is sent back to the user. Note that even
though catching such a generic exception (as Exeception for possible non
recoverable issues) is not usually good, in this case is better to
report some severe message to the user (and still print the stack trace)
rather than letting the GUI reload wheel spinning indefinitely (as it was
often the case)...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>