app: js: synchrona: update input reference status led
With commit 72f9b79a12d5
("app: python: synchrona: update input reference report"), the app now
reports when the ad9545 enters in holdover mode in which case we print an
'alert()' message and use the 'warning' LED status.

On top of this, when the reference being used is OCXO, we change the
status led so that it flashes green. Note that there is no fundamental
difference between when holdover or free_run (reported as OCXO) mode is
being used. The only reason we use a 'warning' led in holdover is to notify
that we had a valid input reference that was lost. On the other hand,
when OCXO is reported, it means we never had a valid input reference but
the device is still capable of working in "standalone" mode (as it is in
holdover mode).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>