app: js: synchrona: update input reference status
Update the input references on the block diagram according with their
availability. Note that for REF_IN and PPS, we can always know if
there is a valid input since the ad9545 chip can report that. For the
other references we can only know if we have a valid one if it's being
used and the PLLs are locked.

On top of this, TCXO is now only show as "enabled" in the diagram if
it's, in fact, being used. Hence 'setVCXO_TCXO()' was renamed and
updated accordingly.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>