app: js: synchrona: improve PPL2 frequency calculation
The calculation of the PLL2 frequency was just getting channel 0 and
multiply it's frequency with it's divider. However, if the channel is
not enabled, that calculation won't exactly match the PLL2 frequency due
to integer approximations. Most likely, the approximated value won't really
have any side effect on the coarse delay calculation but better not to risk
it and do things properly. Thus, we find out the first *enabled* channel
and then calculate the PLL2 frequency.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>