config: dhcpcd.conf: fix default settings
The regex executed to remove interface entries is expecting an empty
line between each entry otherwise the executed pattern won't match and
the entry is not removed. Hence, add a new line in the end of the file
so that if the user want's to change to dynmaic ip, the entry is
removed.

Note that this is the first part to fix dynamic ip configuration.
A following patch will complete the fix on the php/html side.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>