included: dhcp: fix dynamic ip configuration
After configuring a static ip, the only way to change back to dynamic ip
was to also add dnsmask settings. This should not be necessary as the
user might want to change between dynamic and static ip without having
to set a dhcp server on the interface (when static ip is used).

To solve the issue, we add a different value ("dhcp") to the DHCP radio
button so that we can distinguish between dynamic vs static ip on the
target side and act accordingly.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>